separate css add magnifying glass to search page refactor loading to logo improve tests add Sass add css source maps
