+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "alt": "^0.17.1",
     "assert": "^1.3.0",
+    "autoprefixer-loader": "^3.1.0",
     "babel": "^5.8.21",
     "babel-core": "^5.8.22",
     "babel-loader": "^5.3.2",
@@ -31,16 +32,19 @@
     "compression": "^1.5.2",
     "connect-redis": "^2.5.0",
     "cookie-parser": "^1.3.5",
+    "css-loader": "^0.18.0",
     "datamaps": "^0.4.0",
     "elasticsearch": "^8.0.1",
     "express": "^4.13.3",
     "express-handlebars": "^2.0.1",
     "express-session": "^1.11.3",
+    "extract-text-webpack-plugin": "^0.8.2",
     "file-loader": "^0.8.4",
     "iso": "^4.2.0",
     "mocha": "^2.2.5",
     "moment": "^2.10.6",
     "moment-range": "^2.0.3",
+    "node-sass": "^3.3.3",
     "nodemon": "^1.4.1",
     "object-assign": "^4.0.1",
     "passport": "^0.3.0",
@@ -55,6 +59,8 @@
     "react-hot-loader": "^1.2.8",
     "react-router": "^0.13.3",
     "redis": "^1.0.0",
+    "sass-loader": "^2.0.1",
+    "style-loader": "^0.12.4",
     "superagent": "^1.3.0",
     "testdom": "^1.1.0",
     "underscore": "^1.8.3",
