+++ src/client/styles/_logo.scss
@@ -1,0 +1,291 @@
+.logo-container{
+  position: relative;
+  height: 250px;
+  width: 250px;
+  margin: 0 auto;
+  font-size:8px; /* used for border width */
+}
+.logo--hide{
+  opacity:0!important;
+  animation:none!important;
+}
+
+.logo-container > *{
+  height: 100%;
+  width: 100% ;
+}
+
+[class*=lantern__],
+.lantern__glass::before,
+.lantern__glass::after,
+.lantern__base::after{
+  content: '';
+  display: block;
+  border: 0.75em solid black;
+  margin: 0 auto;
+}
+.lantern__circle{
+  height: 100%; width: 100%;
+  /*border: 1em solid black;*/
+  border: transparent;
+  border-radius:50%;
+}
+.lantern__hoop{
+  height: 16%;
+  width: 16%;
+  border-radius:50%;
+}
+.lantern__base--lower,
+.lantern__lid--lower,
+.lantern__lid,
+.lantern__glass,
+.lantern__neck,
+.lantern__base{
+  width:39%;
+  height:7.5%;
+  border-radius:0.75em 0.75em 0 0;
+  margin-top: -0.75em;
+}
+.lantern__glass{
+  position: relative;
+  height: 40.5%;
+  border-radius:0 0 2.5em 2.5em;
+}
+.lantern__flame{
+  width: 45.44%;
+  height: 43.5%;
+  border-bottom-right-radius: 50%;
+  border-bottom-left-radius: 50%;
+  border-top-left-radius: 50%;
+  transform:rotate(-45deg);
+  border-width: 0.4em;
+  position: absolute;
+  bottom: 12.1%;
+  left: calc(50% - 22.5%);
+  z-index: 1;
+  background-color: white;
+}
+.lantern__neck{
+  width: calc(50% - 7.5em);
+}
+.lantern__base{
+  height: 21.5%;
+  border-radius:2.5em 2.5em 0 0;
+}
+.lantern__base::after{
+  height: 0;
+  width: 0;
+  margin: 20% auto;
+  border-radius: 0.75em;
+}
+.lantern__base--lower,
+.lantern__lid--lower{
+  width:50%;
+  border-radius: 0.75em;
+  height:8.55%;
+}
+.lantern__glass::after,
+.lantern__glass::before{
+  opacity: .5;
+  position: relative;
+  left: 25%;
+  top:1.25em;
+  height:25%;
+  width:2px;
+  border-width:2px;
+  border-radius: 0.3em;
+  border-color: transparent;
+}
+.lantern__glass::after{
+  top:-1.3em;
+  left: 33.3%;
+  height:50%;
+}
+.lantern__base--lower::after{
+  display: none;
+}
+.lantern__wick{
+  width: 1px;
+  height: 13.5%;
+  border-width: 0.25em;
+  position: absolute;
+  bottom: 0;
+  left: calc(50% - 0.2em);
+  transform-origin: bottom;
+}
+.loading .lantern__glass::after,
+.loading .lantern__glass::before {
+  animation: flicker-reflection 3s backwards linear infinite;
+}
+.loading .lantern__wick {
+  animation: flicker-wick 3s backwards linear infinite;
+}
+.loading .lantern__flame {
+  background-image: radial-gradient(59% 59%, red, orange);
+  animation: flicker 3s backwards  linear infinite;
+}
+
+.error .lantern__flame{
+  width: 1.25em;
+  opacity: 0.35;
+  height: 60px;
+  transform: rotate(0deg) translateX(10px);
+  background-image: linear-gradient(white, rgba(255, 255, 255, 0));
+  background-color: black;
+  border: 0;
+  animation: flicker-out 2s forwards alternate infinite;
+  left: calc(50% - 1.9em);
+}
+
+@keyframes flicker-out {
+  0% {
+    background-color: black;
+  }
+  100% {
+    background-color: lightgrey;
+  }
+}
+
+@keyframes flicker-reflection {
+  5% {
+    border-color:yellow;
+  }
+  12% {
+    border-color:orange;
+  }
+  38% {
+    border-color:orange;
+  }
+  75.5% {
+    border-color:yellow;
+  }
+  85.5% {
+    border-color:orange;
+  }
+  100% {
+    border-color:yellow;
+  }
+}
+
+@keyframes flicker-wick {
+  12% {
+    transform: rotate(10deg) translateY(0) translateX(0px);
+  }
+  65.5%  {
+    transform: rotate(-5deg) translateY(0) translateX(0px);
+  }
+  75.5% {
+    transform: rotate(0deg) translateY(0) translateX(0px);
+  }
+  85.5% {
+    transform: rotate(10deg) translateY(0) translateX(0px);
+  }
+}
+
+@keyframes flicker {
+  5% {
+    transform: rotate(-50deg);
+  }
+  12% {
+  transform: rotate(-35deg);
+    height: 47.5%;
+    width: 47.5%;
+  }
+  38% {
+  transform: rotate(-32deg);
+    height: 45%;
+    width: 52.5%;
+  }
+  75.5% {
+  transform: rotate(-55deg) ;
+    height: 43.5%;
+    width: 42.5%;
+  }
+  85.5% {
+  transform: rotate(-50deg) ;
+    height: 45%;
+    width: 55.44%;
+  }
+  100% {
+    transform: rotate(-45deg);
+  }
+
+}
+
+.lantern--search .lantern__circle{
+  border: transparent;
+}
+.lantern--search.lantern{
+  transform:scale(1.15);
+  opacity: 0.85;
+}
+body.loading-mounted .lantern {
+  animation:lantern--load 2s ease-in-out forwards;
+}
+body.loading-mounted .magnifying-glass {
+  animation:magnifying-glass--load 2s ease-in-out forwards;
+}
+
+.magnifying-glass__join,
+.magnifying-glass__rod{
+  content: '';
+  display: block;
+  border: 0.75em solid black;
+  margin: 0 auto;
+}
+
+.magnifying-glass{
+  transform: rotate(-60deg);
+  border:1em solid black;
+  position: absolute; top: -0px;
+  border-radius: 50%;
+  z-index: 2;
+
+  width: 120%;
+  height: 120%;
+  top: -10%;
+  left: -10%;
+}
+
+.magnifying-glass__glass{
+  background: radial-gradient(circle at bottom,#81e8f6,#76deef 10%,#055194 80%,#062745 100%);
+  opacity: 0.2;
+  display: block; width: 100%; height:100%;
+  border-radius: 50%;
+}
+.magnifying-glass__glass:after{
+  display: block;
+  background: radial-gradient(circle at 50% 80%,rgba(255,255,255,0),rgba(255,255,255,0) 74%,white 80%,white 84%,rgba(255,255,255,0) 100%);
+  content: "";
+  position: absolute;
+  top: 4.5%;
+  left: 4.5%;
+  width: 90%;
+  height: 90%;
+  border-radius: 100%;
+  filter: blur(1px);
+  z-index: 2;
+}
+.magnifying-glass__join{
+  width:20%;
+  height:7%;
+}
+.magnifying-glass__rod{
+  width:15%;
+  height:40%;
+  margin-top:-0.75em;
+  position: relative;
+}
+
+@keyframes magnifying-glass--load {
+  100%{
+    transform: scale(1.15) rotate(-60deg);
+    opacity: 0;
+  }
+}
+@keyframes lantern--load {
+  100%{
+    transform:scale(1);
+    opacity: 1;
+  }
+}
