+++ src/server/views/index.hbs
@@ -1,4 +1,4 @@
 <div id="react-app">
   {{{content}}}
 </div>
-<script src="{{jsUrl}}/js/app.js{{version}}" defer></script>
+<script src="{{jsUrl}}/build/scripts/app.js{{version}}" defer></script>
