+++ src/server/views/layouts/main.hbs
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1"/>
     <link href='https://fonts.googleapis.com/css?family=Work+Sans' rel='stylesheet' type='text/css'>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">
-    <link rel="stylesheet" href="/app.css">
+    <link rel="stylesheet" href="{{jsUrl}}/build/styles/app.css{{version}}">
 
     <link rel="apple-touch-icon-precomposed" sizes="57x57" href="/favicons/apple-icon-57x57.png" />
     <link rel="apple-touch-icon-precomposed" sizes="114x114" href="/favicons/apple-icon-114x114.png" />
